% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize.r
\name{arcpy.initialize}
\alias{arcpy.initialize}
\title{Initialize the arcypy interface}
\usage{
arcpy.initialize(PYTHON_EXE, quietly = FALSE)
}
\arguments{
\item{PYTHON_EXE}{Path to python executable. If missing, will search for the
default Windows path to the ArcGIS Python distribution.}

\item{quietly}{Logical: Suppress messages and warnings.}
}
\value{
No return value.
}
\description{
Initialize the R-arcypy interface by opening a connection to the
  ArcGIS Python installation, importing arcpy, and generating
  interfaces for arcpy functions and objects.
}
\details{
The following arcpy calls are available:
\subsection{Environment Settings}{
  \itemize{
    \item env.workspace
    \item env.overwriteOutput
  }
}
\subsection{Listing}{
  \itemize{
    \item ListDatasets
    \item ListFeatureClasses
    \item ListFields
    \item ListFiles
    \item ListIndexes
    \item ListRasters
    \item ListTables
    \item ListVersions
    \item ListWorkspaces
  }
}
\subsection{Extraction}{
  \itemize{
    \item Clip_analysis
    \item Select_analysis
  }
}
\subsection{Management}{
  \itemize{
    \item JoinField_management
  }
}
\subsection{Conversion}{
  \itemize{
    \item RasterToPolygon_conversion
  }
}
}
\examples{
\dontrun{
arcpy.initialize("C:/Python27/ArcGIS10.3/python.exe")
env.workspace("C:/test.gdb")
env.overwriteOutput(TRUE)
Clip_analysis("feature1", "feature2")
}
}

