% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{arcpy_env}
\alias{arcpy_env}
\title{ArcPy Environment}
\usage{
arcpy_env(function.list = NULL)
}
\arguments{
\item{function.list}{A list of functions from the module to expose. 
If NULL, all functions in the module will be exposed.}
}
\value{
An environment object containing arcpy functions.
}
\description{
Create an environment that contains ArcPy function interfaces.
}
\examples{
\dontrun{
arcpy = arcpy_env()
ls(arcpy)

arcpy$Delete_management(path_to_file)

arcpy = arcpy_env(list(Delete_management, Clip_anaylsis))
ls(arcpy)
}

}

