% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{attach_custom}
\alias{attach_custom}
\title{Attach Custom Toolbox or Module}
\usage{
attach_custom(envir, name, function.list = NULL, input.file)
}
\arguments{
\item{envir}{The environment to attach the module to.}

\item{name}{The name of the toolbox or module to attach.}

\item{function.list}{A list of functions from the module to expose. 
If NULL, all functions in the module will be exposed.}

\item{input.file}{The input file containing the toolbox.}
}
\description{
Attach a custom toolbox or module to ArcPy.
}

