% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{attach_module}
\alias{attach_arc}
\alias{attach_ba}
\alias{attach_da}
\alias{attach_ddd}
\alias{attach_fabric}
\alias{attach_ga}
\alias{attach_interop}
\alias{attach_mapping}
\alias{attach_module}
\alias{attach_mp}
\alias{attach_na}
\alias{attach_sa}
\alias{attach_schematics}
\alias{attach_ta}
\title{Attach Module}
\usage{
attach_module(envir, module.name, function.list = NULL, checkout = NA)

attach_sa(envir, function.list = NULL)

attach_ga(envir, function.list = NULL)

attach_na(envir, function.list = NULL)

attach_ddd(envir, function.list = NULL)

attach_ta(envir, function.list = NULL)

attach_ba(envir, function.list = NULL)

attach_da(envir, function.list = NULL)

attach_mp(envir, function.list = NULL)

attach_mapping(envir, function.list = NULL)

attach_fabric(envir, function.list = NULL)

attach_arc(envir, function.list = NULL)

attach_interop(envir, function.list = NULL)

attach_schematics(envir, function.list = NULL)
}
\arguments{
\item{envir}{The environment to attach the module to.}

\item{module.name}{The name fo the module to attach.}

\item{function.list}{A list of functions from the module to expose. 
If NULL, all functions in the module will be exposed.}

\item{checkout}{If required, the name of the ArcGIS extension to 
check out.}
}
\value{
(Invisible) The destination environment \code{envir}.
}
\description{
Attach a module to arcpy environment.
}
\details{
Technically, any Python module can be attached.
}
\section{Functions}{
\itemize{
\item \code{attach_sa}: Attach the ArcGIS Spatial Analyst module \code{arcpy.sa}.

\item \code{attach_ga}: Attach the ArcGIS Geostatistical Analyst module \code{arcpy.ga}.

\item \code{attach_na}: Attach the ArcGIS Network Analyst module \code{arcpy.na}.

\item \code{attach_ddd}: Attach the ArcGIS 3D Analyst module \code{arcpy.ddd}.

\item \code{attach_ta}: ArcGIS Tracking Analyst module \code{arcpy.ta}.

\item \code{attach_ba}: Attach the ArcGIS Business Analyst module \code{arcpy.ba}.

\item \code{attach_da}: Attach the ArcGIS Data Access module \code{arcpy.da}.

\item \code{attach_mp}: Attach the ArcGIS Pro Mapping module \code{arcpy.mp}.

\item \code{attach_mapping}: Attach the ArcGIS Desktop mapping module \code{arcpy.mapping}.

\item \code{attach_fabric}: Attach the ArcGIS Desktop Fabric module \code{arcpy.fabric}.

\item \code{attach_arc}: Attach the ArcGIS Desktop Coverage module \code{arcpy.arc}.

\item \code{attach_interop}: Attach the ArcGIS Desktop Data Interoperability module \code{arcpy.interop}.

\item \code{attach_schematics}: Attach the ArcGIS Desktop Schematics module \code{arcpy.schematics}.
}}

