% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{attach_toolbox}
\alias{attach_toolbox}
\title{Attach Toolbox}
\usage{
attach_toolbox(envir, name, tool.list)
}
\arguments{
\item{envir}{The environment to attach the toolbox to.}

\item{name}{The name of the toolbox.}

\item{tool.list}{A list of tools from the toolbox to expose. 
If NULL, all tools in the toolbox will be exposed.}
}
\description{
Attach an ArcGIS toolbox.
}
\seealso{
\code{\link{attach_custom}}
}

