% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{checkout_extension}
\alias{checkout_extension}
\title{Checkout Extension}
\usage{
checkout_extension(ext)
}
\arguments{
\item{ext}{The extension to check out.}
}
\description{
Checkout an ArcGIS extension.
}
