% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{RasterCalculator}
\alias{RasterCalculator}
\title{Raster Calculation with arcpyr}
\usage{
RasterCalculator(expressions, inrasters = list(), outrasters = list())
}
\arguments{
\item{expressions}{A list of raster calculation expressions to be 
executed in sequence.}

\item{inrasters}{A named list of file paths to Rasters used in the 
calculation expressions, to be converted to ArcPy raster objects.}

\item{outrasters}{A named list of file paths to save Rasters 
resulting the calculation expressions.}
}
\value{
No return value; raster files listed in \code{outrasters} 
  will be saved.
}
\description{
Raster Calculation with arcpy is handled via this helper function.
}

