% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{tool_help}
\alias{tool_help}
\title{R interface to arcpy help}
\usage{
tool_help(tool)
}
\arguments{
\item{tool}{The tool to get help for.}
}
\description{
Print the help documentation for an arcpy function.
}
\examples{
\dontrun{
arcpy = arcpy_env()
tool_help(arcpy$Delete_managment)
tool_help(arcpy$env)
}
}

