% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize.r
\name{arcpy.initialize}
\alias{arcpy.initialize}
\title{Initialize the arcypy interface}
\usage{
arcpy.initialize(PYTHON_EXE)
}
\description{
Initialize the R-arcypy interface by opening a connection to the
  ArcGIS Python installation, importing arcpy, and generating
  interfaces for arcpy functions and objects.
}
\details{
The following arcpy calls are available:
  \itemize{
    \item \strong{Environment Settings}
    \item env.workspace
    \item env.overwriteOutput
    \item \strong{Listing}
    \item ListDatasets
    \item ListFeatureClasses
    \item ListFields
    \item ListFiles
    \item ListIndexes
    \item ListRasters
    \item ListTables
    \item ListVersions
    \item ListWorkspaces
    \item \strong{Extraction}
    \item Clip_analysis
    \item Select_analysis
  }
}

