% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{arcpy.initialize}
\alias{arcpy.initialize}
\title{Initialize the arcypy interface}
\usage{
arcpy.initialize(PYTHON_EXE, quietly = FALSE)
}
\arguments{
\item{PYTHON_EXE}{Path to python executable. If missing, will search for the
default Windows path to the ArcGIS Python distribution.}

\item{quietly}{Logical: Suppress messages and warnings.}
}
\value{
No return value.
}
\description{
Initialize the R-arcypy interface by opening a connection to the 
  ArcGIS Python installation, importing arcpy, and generating 
  interfaces for arcpy functions and objects.
}
\details{
The following arcpy calls are available:
\subsection{Environment Settings}{
  \itemize{
    \item env.workspace
    \item env.overwriteOutput
  }
}
\subsection{Listing}{
  \itemize{
    \item ListDatasets 
    \item ListFeatureClasses 
    \item ListFields 
    \item ListFiles 
    \item ListIndexes 
    \item ListRasters 
    \item ListTables 
    \item ListVersions 
    \item ListWorkspaces
  }
}
\subsection{Analysis}{
  \itemize{
    \item Clip_analysis 
    \item Select_analysis
    \item Intersect_analysis
    \item SpatialJoin_analysis
    \item Union_analysis
    \item Identity_analysis
  }
}
\subsection{Management}{
  \itemize{
    \item AddJoin_management
    \item RemoveJoin_management
    \item JoinField_management
    \item Dissolve_management
    \item Delete_management
    \item MakeFeatureLayer_management
    \item SelectLayerByAttribute_management
    \item SelectLayerByLocation_management
    \item AlterField_management
    \item CopyFeatures_management
    \item AddField_management
    \item CalculateField_management
    \item DeleteField_management
    \item FeatureToPoint_management
    \item Merge_management
    \item Append_management
  }
}
\subsection{Conversion}{
  \itemize{
    \item RasterToPolygon_conversion
    \item TableToTable_conversion
  }
}
}
\examples{
\dontrun{
arcpy.initialize("C:/Python27/ArcGIS10.3/python.exe")
env.workspace("C:/test.gdb")
env.overwriteOutput(TRUE)
Clip_analysis("feature1", "feature2")
}

}

