% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildenv.r
\name{attach_custom}
\alias{attach_custom}
\title{Attach a Custom Toolbox}
\usage{
attach_custom(envir, name, tool.list, input.file)
}
\arguments{
\item{envir}{The environment to attach the toolbox to.}

\item{name}{The alias of the toolbox to attach. 
Underscores \code{_} are not allowed.}

\item{tool.list}{A list of tools from the toolbox to expose. 
If NULL, all tools in the toolbox will be exposed.}

\item{input.file}{The input file containing the toolbox.}
}
\description{
Attach a custom toolbox to ArcPy.
}
\seealso{
\code{\link{attach_toolbox}}
}
