% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize.r
\name{sa.initialize}
\alias{sa.initialize}
\title{Initialize Spatial Analyst Tools}
\usage{
sa.initialize()
}
\value{
No return value.
}
\description{
Initialize functions from the Spatial Analyst module.
}
\details{
The following arcpy.sa calls are available:
\subsection{Raster tools}{
  \itemize{
    \item RasterCalculator
  }
}
}
\examples{
\dontrun{
sa.initialize()
RasterCalculator(expr, "outrast")
}
}

